requirements.txt
# ── Core ML ────────────────────────────────────────────────────────────────────
numpy>=1.24.0
pandas>=2.0.0
torch>=2.1.0
transformers>=4.35.0
scikit-learn>=1.3.0

# ── Bio ────────────────────────────────────────────────────────────────────────
biopython>=1.81

# ── Visualization ─────────────────────────────────────────────────────────────
matplotlib>=3.7.0
plotly>=5.17.0
seaborn>=0.12.0

# ── Graph Neural Networks ─────────────────────────────────────────────────────
networkx>=3.1
torch-geometric>=2.4.0

# ── Numerical / Scientific ────────────────────────────────────────────────────
scipy>=1.11.0

# ── Web / API ─────────────────────────────────────────────────────────────────
requests>=2.31.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.0.0

# ── Dashboard ─────────────────────────────────────────────────────────────────
dash>=2.14.0
dash-bootstrap-components>=1.5.0

# ── Utilities ─────────────────────────────────────────────────────────────────
tqdm>=4.66.0